Multi render support #137

+++ media/samples/shaders/sources/common/structs_inc.glsl
@@ -4,8 +4,8 @@ struct EnvironmentProbeStruct
 {
     vec4 position;      // Probe position
     mat4 projView[6];   // View-Projection matrix for every cube face
-    int flags;
+    uint flags;
 };
 
-#define ENV_PROBE_FLAG_IRRADIANCE   1
-#define ENV_PROBE_FLAG_SPECULAR     2
+#define ENV_PROBE_FLAG_IRRADIANCE   uint(1)
+#define ENV_PROBE_FLAG_SPECULAR     uint(2)
